[addons][pvr] move all "C" parts to independent header place
This thought to have it cleaner and to take for other languages.
