+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/c-api/addon-instance/CMakeLists.txt
@@ -1,4 +1,5 @@
-set(HEADERS image_decoder.h)
+set(HEADERS image_decoder.h
+            pvr.h)
 
 if(NOT ENABLE_STATIC_LIBS)
   core_add_library(addons_kodi-addon-dev-kit_include_kodi_c-api_addon-instance)
