+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/c-api/addon-instance/pvr/CMakeLists.txt
@@ -1,0 +1,15 @@
+set(HEADERS pvr.h
+            pvr_channel_groups.h
+            pvr_channels.h
+            pvr_defines.h
+            pvr_edl.h
+            pvr_epg.h
+            pvr_general.h
+            pvr_menu_hook.h
+            pvr_recordings.h
+            pvr_stream.h
+            pvr_timers.h)
+
+if(NOT ENABLE_STATIC_LIBS)
+  core_add_library(addons_kodi-addon-dev-kit_include_kodi_c-api_addon-instance_pvr)
+endif()
