+++ xbmc/addons/kodi-addon-dev-kit/include/kodi/c-api/addon-instance/pvr/pvr_channel_groups.h
@@ -1,0 +1,38 @@
+/*
+ *  Copyright (C) 2005-2018 Team Kodi
+ *  This file is part of Kodi - https://kodi.tv
+ *
+ *  SPDX-License-Identifier: GPL-2.0-or-later
+ *  See LICENSES/README.md for more information.
+ */
+
+#pragma once
+
+#include "pvr_defines.h"
+
+#include <stdbool.h>
+
+#ifdef __cplusplus
+extern "C"
+{
+#endif /* __cplusplus */
+
+  typedef struct PVR_CHANNEL_GROUP
+  {
+    char strGroupName[PVR_ADDON_NAME_STRING_LENGTH];
+    bool bIsRadio;
+    unsigned int iPosition;
+  } PVR_CHANNEL_GROUP;
+
+  typedef struct PVR_CHANNEL_GROUP_MEMBER
+  {
+    char strGroupName[PVR_ADDON_NAME_STRING_LENGTH];
+    unsigned int iChannelUniqueId;
+    unsigned int iChannelNumber;
+    unsigned int iSubChannelNumber;
+    int iOrder;
+  } PVR_CHANNEL_GROUP_MEMBER;
+
+#ifdef __cplusplus
+}
+#endif /* __cplusplus */
